+++ griptape/schemas/base_schema.py
@@ -208,7 +208,6 @@ class BaseSchema(Schema):
                 "Sequence": Sequence,
                 "TaskMemory": TaskMemory,
                 "State": BaseTask.State,
-                "ConversationMemoryStrategy": Structure.ConversationMemoryStrategy,
                 "BaseConversationMemory": BaseConversationMemory,
                 "BaseArtifactStorage": BaseArtifactStorage,
                 # Third party modules