+++ griptape/structures/structure.py
@@ -2,8 +2,7 @@ from __future__ import annotations
 
 import uuid
 from abc import ABC, abstractmethod
-from enum import Enum
-from typing import TYPE_CHECKING, Any, Optional, Union
+from typing import TYPE_CHECKING, Any, Literal, Optional, Union
 
 from attrs import Factory, define, field
 
@@ -24,10 +23,6 @@ if TYPE_CHECKING:
 
 @define
 class Structure(RuleMixin, SerializableMixin, RunnableMixin["Structure"], ABC):
-    class ConversationMemoryStrategy(Enum):
-        PER_STRUCTURE = 1
-        PER_TASK = 2
-
     id: str = field(default=Factory(lambda: uuid.uuid4().hex), kw_only=True, metadata={"serializable": True})
     _tasks: list[Union[BaseTask, list[BaseTask]]] = field(
         factory=list, kw_only=True, alias="tasks", metadata={"serializable": True}
@@ -37,8 +32,8 @@ class Structure(RuleMixin, SerializableMixin, RunnableMixin["Structure"], ABC):
         kw_only=True,
         metadata={"serializable": True},
     )
-    conversation_memory_strategy: ConversationMemoryStrategy = field(
-        default=ConversationMemoryStrategy.PER_STRUCTURE, kw_only=True, metadata={"serializable": True}
+    conversation_memory_strategy: Literal["per_structure", "per_task"] = field(
+        default="per_structure", kw_only=True, metadata={"serializable": True}
     )
     task_memory: TaskMemory = field(
         default=Factory(lambda self: TaskMemory(), takes_self=True),
@@ -172,7 +167,7 @@ class Structure(RuleMixin, SerializableMixin, RunnableMixin["Structure"], ABC):
 
         if self.output_task is not None:
             if (
-                self.conversation_memory_strategy == self.ConversationMemoryStrategy.PER_STRUCTURE
+                self.conversation_memory_strategy == "per_structure"
                 and self.conversation_memory is not None
                 and self.input_task is not None
                 and self.output_task.output is not None